Some slight changes

+++ Quick English Vietnamese Dictionary/Classes/ExternalLibs/SVProgressHud/SVProgressHUD.m
@@ -41,7 +41,7 @@ CGFloat SVProgressHUDRingThickness = 6;
 
 @property (nonatomic, readonly) CGFloat visibleKeyboardHeight;
 
-- (void)showProgress:(float)progress
+- (void)showProgress:(CGFloat)progress
               status:(NSString*)string
             maskType:(SVProgressHUDMaskType)hudMaskType;
 
@@ -398,7 +398,7 @@ CGFloat SVProgressHUDRingThickness = 6;
 
 #pragma mark - Master show/dismiss methods
 
-- (void)showProgress:(float)progress status:(NSString*)string maskType:(SVProgressHUDMaskType)hudMaskType {
+- (void)showProgress:(CGFloat)progress status:(NSString*)string maskType:(SVProgressHUDMaskType)hudMaskType {
     
     if(!self.overlayView.superview){
         NSEnumerator *frontToBackWindows = [[[UIApplication sharedApplication]windows]reverseObjectEnumerator];
